data_analysis_tools/requirements.txt
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
fire==0.5.0
fonttools==4.43.0
h5py==3.9.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.0
mdurl==0.1.2
natsort==8.4.0
nptyping==2.5.0
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
Pygments==2.16.1
pynrrd==1.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
rich==13.5.3
scipy==1.11.3
six==1.16.0
termcolor==2.3.0
tqdm==4.66.1
typing_extensions==4.8.0